Loading and Error handling moved to MoviePage/index.js
