+++ src/features/MoviePage/styled.js
@@ -1,5 +1,9 @@
 import styled from "styled-components";
 
+export const LoadingPage = styled.div``;
+
+export const ErrorPage = styled.div``;
+
 export const Container = styled.div`
   margin: 0 auto;
   margin-bottom: 336px;
